+++ packages/adguard-api/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2023-08-18
+
+### Changed
+- Update `@adguard/tswebextension` to `v0.3.10`
+  
+
 ## [1.3.0] - 2023-08-10
 ### Changed
 - [BREAKING CHANGE] Assistant script is separated and dynamically injected. You need to provide it in your project.