+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
-## UNRELEASED
+## [Unreleased]
 
 ## Added
 - New `$permissions` modifier to set Permissions-Policy response header.
@@ -16,6 +16,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support conversion to DNR for `$permissions` modifier.
 - Support conversion to DNR for `$cookie` modifier.
 - `$url` modifier support for non-basic rules.
+
+
+## [2.1.9] - 2023-08-18
+
+### Added
+- Added `CspReport` to RequestType enum.
+
+
+## [2.1.8] - 2023-08-14
+
+### Fixed
+- Crash on converting allowlist csp rules.
 
 
 ## [2.1.7] - 2023-08-10