+++ packages/tswebextension/CHANGELOG.md
@@ -7,13 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
-## UNRELEASED
+## [Unreleased]
 
 ## Fixed
 - Applying cookie rules in content-script in MV2.
 
 ## Added
 - Support for `$cookie` modifier in MV3 via browser.cookies API and content-script.
+
+
+## [0.3.10] - 2023-08-18
+
+### Added
+- Blocking third-party requests with `csp_report` content-type.
+- Handling discarded tabs replacement on wake up.
+
+### Fixed
+- Do not expose JS rules in global page scope.
 
 
 ## [0.3.9] - 2023-08-10